[HC-194] CourseReviews 1차 리팩토링 완료
