+++ src/constants/semesters.ts
@@ -1,0 +1,21 @@
+export const semesters = [
+    '17-18년도 1학기',
+    '17-18년도 2학기',
+    '17-18년도 3학기/계절학기',
+    '18-19년도 1학기',
+    '18-19년도 2학기',
+    '18-19년도 3학기/계절학기',
+    '19-20년도 1학기',
+    '19-20년도 2학기',
+    '19-20년도 3학기/계절학기',
+    '20-21년도 1학기',
+    '20-21년도 2학기',
+    '20-21년도 3학기/계절학기',
+    '21-22년도 1학기',
+    '21-22년도 2학기',
+    '21-22년도 3학기/계절학기',
+    '22-23년도 1학기',
+    '22-23년도 2학기',
+    '22-23년도 3학기/계절학기',
+    '23-24년도 1학기',
+];
