+++ src/index.css
@@ -24,9 +24,3 @@ body {
 .nav-link:enabled {
   color: black !important;
 }
-
-.btn-primary:enabled {
-  background-color: #43a680 !important;
-  border-color: #43a680 !important;
-  color: white !important;
-}
