+++ src/pages/CourseReviews/CourseReviews.tsx
@@ -10,6 +10,7 @@ import { apiDelete, apiGet, apiPut } from "../../API/APIHandler";
 import { useReviews } from "../../API/reviews/useReviews";
 import Loader from "../../components/Loader/Loader";
 import { useEditReview } from "../../API/reviews/useEditReview";
+import ReviewCard from "./components/ReviewCard/ReviewCard";
 
 // 수업 리뷰 디스플레이 컴포넌트 (https://honeycourses.com/course/view/수업ID)
 
@@ -30,22 +31,8 @@ interface Review {
 }
 
 export default function CourseReviews() {
-  const [editingReviewId, setEditingReviewId] = useState<number | null>(null);
-  const [editedTitle, setEditedTitle] = useState("");
-  const [editedContent, setEditedContent] = useState("");
-  const [editedInstructor, setEditedInstructor] = useState("");
-  const [editedSemyr, setEditedSemyr] = useState("");
-  const [editedGrade, setEditedGrade] = useState("");
-  const [isEditing, setIsEditing] = useState(false);
   const courseId = window.location.pathname.split("/").pop();
   const navigate = useNavigate();
-  const [isShowMore, setShowMore] = useState(false);
-  const [showMoreReviewId, setShowMoreReviewId] = useState<number | null>(null);
-  const { editSingleReview } = useEditReview();
-
-  const dateA = new Date("2022/06/01 08:00:00");
-  const dateB = new Date("2022/06/01 00:00:00");
-  const diffMSec = dateA.getTime() - dateB.getTime();
 
   /*const handleUpvote = async (reviewId: number) => {
     try {
@@ -80,6 +67,10 @@ export default function CourseReviews() {
     }
   };
 
+  const addReviewClick = () => {
+    navigate(`/courses/addReview/${courseId}`);
+  }
+
   const { isLoading, error, reviews, course_name } = useReviews(courseId);
 
   useEffect(() => {
@@ -109,7 +100,7 @@ export default function CourseReviews() {
 
           <Button
             className="my-auto"
-            href={`/courses/addReview/${courseId}`}
+            onClick={addReviewClick}
             variant="success"
             size="sm"
             style={{
@@ -134,7 +125,7 @@ export default function CourseReviews() {
           <div style={{ textAlign: "center", marginTop: "50px" }}>
             <h3>아직 작성된 리뷰가 없습니다.</h3>
             <Button
-              href={`/courses/addReview/${courseId}`}
+              onClick={addReviewClick}
               variant="success"
               size="sm"
               style={{ marginTop: "20px" }}
@@ -144,282 +135,7 @@ export default function CourseReviews() {
           </div>
         ) : (
           reviews.map((review: Review) => (
-            <Card key={review.review_id} className={styles.reviewCard}>
-              {review.mine && (
-                <div>
-                  {editingReviewId === review.review_id ? (
-                    <div className={styles.editButtons}>
-                      <Button
-                        variant="primary"
-                        onClick={() => {
-                          editSingleReview({
-                            reviewId: review.review_id,
-                            courseId,
-                            editedTitle,
-                            editedContent,
-                            editedInstructor,
-                            editedSemyr,
-                            editedGrade,
-                          });
-                          setEditingReviewId(null);
-                          setIsEditing(false);
-                          setShowMore(!isShowMore);
-                          setShowMoreReviewId(review.review_id); // 리뷰 수정 후 리뷰 더보기 버튼 수납
-                        }}
-                        disabled={isEditing}
-                      >
-                        제출
-                      </Button>
-                      <Button
-                        variant="secondary"
-                        onClick={() => setEditingReviewId(null)}
-                      >
-                        취소
-                      </Button>
-                    </div>
-                  ) : (
-                    <div
-                      className={styles.reviewButtons}
-                      style={{
-                        position: "absolute",
-                        top: "10px",
-                        right: "10px",
-                      }}
-                    >
-                      <Button
-                        className={styles.showMoreButton}
-                        style={{
-                          backgroundColor: "transparent",
-                          border: "none",
-                        }}
-                        onClick={() => {
-                          setShowMore(!isShowMore);
-                          setShowMoreReviewId(review.review_id);
-                        }}
-                      >
-                        <img
-                          src="/images/showMoreButton.png"
-                          alt="show-more-icon"
-                        />
-                      </Button>
-                      <div
-                        className={
-                          isShowMore && showMoreReviewId === review.review_id
-                            ? styles.reviewChangeDelete
-                            : styles.hiddenReviewChangeDelete
-                        }
-                      >
-                        <Button
-                          className={styles.ReviewChange}
-                          variant="success"
-                          onClick={() => {
-                            setEditingReviewId(review.review_id);
-                            setEditedTitle(review.review_title);
-                            setEditedContent(review.review_content);
-                            setEditedSemyr(review.taken_semyr);
-                            setEditedInstructor(review.instructor_name);
-                            setEditedGrade(review.grade);
-                          }}
-                        >
-                          수정
-                        </Button>
-                        <Button
-                          className={styles.reviewDelete}
-                          variant="danger"
-                          onClick={() => handleDeleteReview(review.review_id)}
-                        >
-                          삭제
-                        </Button>
-                      </div>
-                    </div>
-                  )}
-                </div>
-              )}
-              <Card.Body className="text-start">
-                {editingReviewId === review.review_id ? (
-                  <>
-                    <Form.Control
-                      type="text"
-                      value={editedTitle}
-                      onChange={(e) => setEditedTitle(e.target.value)}
-                      placeholder="제목"
-                      className="mb-2"
-                      required
-                    />
-                    <Form.Control
-                      type="text"
-                      value={editedInstructor}
-                      onChange={(e) => setEditedInstructor(e.target.value)}
-                      placeholder="교수"
-                      className="mb-2"
-                      required
-                    />
-                    <Form.Select
-                      value={editedSemyr}
-                      as="textarea"
-                      rows={10}
-                      onChange={(e) => setEditedSemyr(e.target.value)}
-                      placeholder="수강 학기"
-                      required
-                    >
-                      <option value="">수강 학기 선택</option>
-                      <option value="17-18년도 1학기">17-18년도 1학기</option>
-                      <option value="17-18년도 2학기">17-18년도 2학기</option>
-                      <option value="17-18년도 3학기/계절학기">
-                        17-18년도 3학기/계절학기
-                      </option>
-                      <option value="18-19년도 1학기">18-19년도 1학기</option>
-                      <option value="18-19년도 2학기">18-19년도 2학기</option>
-                      <option value="18-19년도 3학기/계절학기">
-                        18-19년도 3학기/계절학기
-                      </option>
-                      <option value="19-20년도 1학기">19-20년도 1학기</option>
-                      <option value="19-20년도 2학기">19-20년도 2학기</option>
-                      <option value="19-20년도 3학기/계절학기">
-                        19-20년도 3학기/계절학기
-                      </option>
-                      <option value="20-21년도 1학기">20-21년도 1학기</option>
-                      <option value="20-21년도 2학기">20-21년도 2학기</option>
-                      <option value="20-21년도 3학기/계절학기">
-                        20-21년도 3학기/계절학기
-                      </option>
-                      <option value="21-22년도 1학기">21-22년도 1학기</option>
-                      <option value="21-22년도 2학기">21-22년도 2학기</option>
-                      <option value="21-22년도 3학기/계절학기">
-                        21-22년도 3학기/계절학기
-                      </option>
-                      <option value="22-23년도 1학기">22-23년도 1학기</option>
-                      <option value="22-23년도 2학기">22-23년도 2학기</option>
-                      <option value="22-23년도 3학기/계절학기">
-                        22-23년도 3학기/계절학기
-                      </option>
-                      <option value="23-24년도 1학기">23-24년도 1학기</option>
-                    </Form.Select>
-                    <br />
-                    <Form.Control
-                      value={editedContent}
-                      as="textarea"
-                      rows={10}
-                      onChange={(e) => setEditedContent(e.target.value)}
-                      placeholder="내용"
-                      required
-                    ></Form.Control>
-                    <Form.Control
-                      value={editedGrade}
-                      as="textarea"
-                      rows={1}
-                      onChange={(e) => setEditedGrade(e.target.value)}
-                      placeholder="성적"
-                      required
-                    ></Form.Control>
-                    <br />
-                    <div style={{ width: "65%", height: "50px" }}>
-                      {review.mine && (
-                        <div
-                          style={{
-                            position: "absolute",
-                            top: "10px",
-                            right: "10px",
-                          }}
-                        ></div>
-                      )}
-                    </div>
-                  </>
-                ) : (
-                  // Display Mode
-                  <>
-                    <Card.Title
-                      style={{
-                        color: "#43A680",
-                        display: "flex",
-                        alignItems: "center",
-                        width: "80%",
-                      }}
-                    >
-                      {review.review_title}
-                    </Card.Title>
-
-                    <hr className={styles.divider}></hr>
-                    <Card.Text style={{ whiteSpace: "pre-wrap" }}>
-                      <p className="fw-semibold" style={{ color: "grey" }}>
-                        수강학기: {review.taken_semyr}, 교수:{" "}
-                        {review.instructor_name}
-                      </p>
-
-                      {review.review_content.replace(/<br\s*[/]?>/gi, "\n")}
-
-                      <br></br>
-                      <br></br>
-
-                      <span>
-                        <p className="fw-bold" style={{ display: "inline" }}>
-                          성적:{" "}
-                        </p>
-                        {review.grade}
-                      </span>
-                      <br></br>
-                    </Card.Text>
-                    <hr className="divider"></hr>
-                    <div style={{ display: "flex", alignItems: "center" }}>
-                      <div
-                        className={styles.date}
-                        style={{
-                          fontSize: "16px",
-                          opacity: 0.7,
-                          marginRight: "auto",
-                        }}
-                      >
-                        <span className={styles.sharp}>
-                          #{review.review_id} {"  "}
-                        </span>
-                        <span className={styles.date}>
-                          {review.review_time === null ? (
-                            ""
-                          ) : (
-                            <>
-                              {new Date(
-                                new Date(review.review_time).getTime() +
-                                  diffMSec
-                              ).toLocaleDateString()}{" "}
-                              작성
-                              {review.updated ? " (수정됨)" : ""}
-                            </>
-                          )}
-                        </span>
-                      </div>
-                      <div
-                        onClick={() => handleUpvote(review.review_id)}
-                        className={
-                          review.liked ? styles.onLikeButton : styles.likeButton
-                        }
-                        style={{ cursor: "pointer" }}
-                      >
-                        <img
-                          src={
-                            review.liked
-                              ? "/images/likeGreen.svg"
-                              : "/images/likeWhiteSolidBlack.svg"
-                          }
-                          alt="likes-icon"
-                          style={{
-                            marginRight: "5px",
-                            width: "20px",
-                            height: "20px",
-                          }}
-                        />
-                        <span
-                          className={
-                            review.liked ? styles.likeCount : styles.none
-                          }
-                        >
-                          {review.like_count}
-                        </span>
-                      </div>
-                    </div>
-                  </>
-                )}
-              </Card.Body>
-            </Card>
+            <ReviewCard courseId={courseId} review={review}/>
           ))
         )}
       </Container>
